--- client/src/attributes.ts
@@ -1,11 +1,0 @@
-export const USER_ATTRIBUTES = ["id", "name", "profilePicture"];
-
-export const CHATROOM_MIN_ATTRIBUTES = ["id", "type", "name"];
-export const CHATROOM_MAX_ATTRIBUTES = [
-  "id",
-  "type",
-  "name",
-  "roomImage",
-  "members",
-  "messages",
-];
