+++ client/src/pages/ChatRoom/RoomList/RoomList.tsx
@@ -3,14 +3,15 @@ import React, { useMemo } from "react";
 import { RoomSubList } from "./RoomSubList/RoomSubList";
 
 import "./room-list.css";
-import { ChatRoomMin } from "types/ChatRoom";
+
 import { Button } from "components/Button/Button";
 import { ROOM_CHANNEL, ROOM_DM } from "constant";
+import { ChatRoomSidebar } from "../ChatRoomType";
 
 type RoomListProps = {
   onClickListItem: (id: string) => void;
   onClickNewChatRoom: () => void;
-  rooms: ChatRoomMin[];
+  rooms: ChatRoomSidebar[];
   selectedRoomId: string;
 };
